refactor(integrations): neutralize upload-init route shells
