+++ turbo/apps/api/src/signals/routes/integrations-telegram-upload-init.ts
@@ -1,5 +1,5 @@
 import { command } from "ccstate";
-import { integrationsPhoneUploadInitContract } from "@okouai/api-contracts/contracts/integrations";
+import { integrationsTelegramUploadInitContract } from "@okouai/api-contracts/contracts/integrations";
 
 import { env } from "../../lib/env";
 import { sanitizeArtifactFilename } from "../../lib/file-url";
@@ -12,23 +12,24 @@ import type { RouteEntry } from "../route-entry";
 
 const PUT_URL_TTL_SECONDS = 3600;
 
-const init$ = command(async ({ get, set }, signal: AbortSignal) => {
+const initInner$ = command(async ({ get, set }, signal: AbortSignal) => {
   const auth = get(authContext$);
+
   const bodyResult = await get(
-    bodyResultOf(integrationsPhoneUploadInitContract.init),
+    bodyResultOf(integrationsTelegramUploadInitContract.init),
   );
   signal.throwIfAborted();
   if (!bodyResult.ok) {
     return bodyResult.response;
   }
 
-  const body = bodyResult.data;
-  const filename = sanitizeArtifactFilename(body.filename);
+  const { filename, contentType, length } = bodyResult.data;
+  const sanitized = sanitizeArtifactFilename(filename);
   const artifact = await set(
     allocateArtifactObject$,
     {
       userId: auth.userId,
-      filename: body.filename,
+      filename,
     },
     signal,
   );
@@ -38,7 +39,7 @@ const init$ = command(async ({ get, set }, signal: AbortSignal) => {
     generatePresignedPutUrl(
       bucket,
       artifact.key,
-      body.contentType,
+      contentType,
       PUT_URL_TTL_SECONDS,
       { usePublicEndpoint: true, metadata: artifact.metadata },
     ),
@@ -51,17 +52,17 @@ const init$ = command(async ({ get, set }, signal: AbortSignal) => {
       uploadId: artifact.id,
       uploadUrl,
       fileUrl: artifact.url,
-      filename,
-      contentType: body.contentType,
-      size: body.length,
+      filename: sanitized,
+      contentType,
+      size: length,
       ...(uploadHeaders ? { uploadHeaders } : {}),
     },
   };
 });
 
-export const zeroIntegrationsPhoneUploadInitRoutes: readonly RouteEntry[] = [
+export const integrationsTelegramUploadInitRoutes: readonly RouteEntry[] = [
   {
-    route: integrationsPhoneUploadInitContract.init,
-    handler: authRoute({ requiredCapability: "phone:write" }, init$),
+    route: integrationsTelegramUploadInitContract.init,
+    handler: authRoute({ requiredCapability: "telegram:write" }, initInner$),
   },
 ];
